create generic web test
